+++ frontend/src/blog.tsx
@@ -1,7 +1,7 @@
 import { hydrateRoot } from "react-dom/client";
 import { BrowserRouter as Router } from "react-router-dom";
 import "./index.css";
-import { BlogTopContainer } from "./blog/container/page/blogs/Container.tsx";
+import { BlogTopContainer } from "./blog/container/page/Container.tsx";
 
 hydrateRoot(
     document.getElementById("root")!,
