+++ frontend/src/blog/container/page/Container.tsx
@@ -1,7 +1,7 @@
 
-import { BlogTop } from "../../../presentation/page/blogs/BlogTop";
+import { BlogTop } from "../../presentation/page/blogs/BlogTop";
 
-import { useGenerateProps } from "./useGenerateProps";
+import { useGenerateProps } from "./blogs/useGenerateProps";
 
 export const BlogTopContainer = () => {
     const generatedProps = useGenerateProps();
