+++ frontend/src/blog/container/page/widgets/Sidebar/Container.tsx
@@ -1,4 +1,4 @@
-import Sidebar from "../../../../../presentation/Sidebar/Sidebar";
+import Sidebar from "../../../../presentation/Sidebar/Sidebar";
 
 import { useGenerateProps } from "./useGenerateProps";
